Be event

See merge request s05-webmobile1-sub3/S05P13C103!58

+++ backend/src/main/java/com/ssafy/api/request/dto/Product/ProductRegisterPostReq.java
@@ -27,10 +27,10 @@ public class ProductRegisterPostReq {
     @ApiModelProperty(name="상품 판매 여부", example="0")
     private Integer is_sold;
     @ApiModelProperty(name="라이브 방송 여부", example="1")
-    private Integer is_live;
+    private Long is_live;
 
     @Builder
-    public ProductRegisterPostReq(String userId, String title, String description, Long category, Integer price, Integer is_sold, Integer is_live) {
+    public ProductRegisterPostReq(String userId, String title, String description, Long category, Integer price, Integer is_sold, Long is_live) {
         this.userId = userId;
         this.title = title;
         this.description = description;
